Time: 2 ms (24.21%), Space: 9.8 MB (48.61%) - LeetHub

+++ 3498-reverse-degree-of-a-string/3498-reverse-degree-of-a-string.cpp
@@ -4,7 +4,7 @@ public:
         int sum = 0;
 
         for (int i = 0; i < s.length(); i++) {
-            sum += (27 - ((int)s[i] - 'a' + 1)) * (i + 1);
+            sum += (27 - (s[i] - 'a' + 1)) * (i + 1);
         }
         return sum;
     }
